transfer Navigation View to GreetingActivity
