+++ .idea/assetWizardSettings.xml
@@ -23,7 +23,7 @@
                                 <PersistentState>
                                   <option name="values">
                                     <map>
-                                      <entry key="url" value="file:/$USER_HOME$/AppData/Local/Android/Sdk/icons/material/materialicons/save_alt/baseline_save_alt_24.xml" />
+                                      <entry key="url" value="file:/$USER_HOME$/AppData/Local/Android/Sdk/icons/material/materialicons/settings/baseline_settings_24.xml" />
                                     </map>
                                   </option>
                                 </PersistentState>
@@ -33,7 +33,7 @@
                         </option>
                         <option name="values">
                           <map>
-                            <entry key="outputName" value="nav_saved" />
+                            <entry key="outputName" value="baseline_settings_24" />
                             <entry key="sourceFile" value="C:\Users\anouk\AndroidStudioProjects\androidApp" />
                           </map>
                         </option>
