+++ app/src/main/java/com/github/romainlogean/sdp_bootcamp/SettingsFragment.java
@@ -1,0 +1,18 @@
+package com.github.romainlogean.sdp_bootcamp;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+public class SettingsFragment extends Fragment {
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_settings, container, false);
+    }
+}
