+++ app/src/main/res/layout/fragment_settings.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".SettingsFragment">
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/settings_fragment"
+        android:textSize="30sp"
+        android:layout_centerInParent="true"
+        android:textColor="@color/material_dynamic_secondary50"/>
+</RelativeLayout>
